fix: calculate string length using string-width
This improves modeline display when CJK characters are used. However, since
the char-width of a CJK character is always 2, regardless of the actual
displayed width, tray information can only be aligned perfectly when half-width
Latin font is used, or in an terminal.

With most Latin fonts, a CJK character is narrower than 2 units of width, so
there will be a blank area at the right of minibuffer. At least that's better
than being pushed to next line.